Users
Add some basic user menu overall working in eirvideo.
need to retrive user Info to be able to use in the rent class.
With this we can ad a Booking History to the customer .
Well work only with customer Name trough this. as it can be connected with email in future develop
pass is not needed, only Mail & BookingHistory & catalog (active Movies)

Co-Authored-By: Jair130 <100359718+Jair130@users.noreply.github.com>
Co-Authored-By: Carlos Alfonso Salinas Staarthof <73845284+carlosstaarthof@users.noreply.github.com>